+++ backend/src/main/java/mediasoft/ru/backend/CurrencyProvider.java
@@ -2,7 +2,7 @@ package mediasoft.ru.backend;
 
 import lombok.Getter;
 import lombok.Setter;
-import mediasoft.ru.backend.product.models.entities.CurrencyEnum;
+import mediasoft.ru.backend.enums.CurrencyEnum;
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.annotation.SessionScope;
