+++ backend/src/main/java/mediasoft/ru/backend/HeaderCaptureFilter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import mediasoft.ru.backend.product.models.entities.CurrencyEnum;
+import mediasoft.ru.backend.enums.CurrencyEnum;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
