+++ backend/src/main/java/mediasoft/ru/backend/models/dto/response/product/ProductInfoResponseDTO.java
@@ -1,4 +1,4 @@
-package mediasoft.ru.backend.dto.response.product;
+package mediasoft.ru.backend.models.dto.response.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,7 @@ import mediasoft.ru.backend.enums.CurrencyEnum;
 import mediasoft.ru.backend.enums.ProductCategory;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -17,7 +18,7 @@ import java.util.UUID;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateProductResponseDTO {
+public class ProductInfoResponseDTO {
     private UUID id;
     private String article;
     private String name;
@@ -25,7 +26,7 @@ public class UpdateProductResponseDTO {
     private ProductCategory category;
     private BigDecimal price;
     private BigDecimal count;
+    private LocalDate creationDate;
     private LocalDateTime lastModifiedDate;
     private CurrencyEnum currency;
 }
-
