+++ backend/src/main/java/mediasoft/ru/backend/services/product/ProductServiceImpl.java
@@ -3,16 +3,16 @@ package mediasoft.ru.backend.services.product;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import mediasoft.ru.backend.CurrencyProvider;
-import mediasoft.ru.backend.dto.ProductDTO;
-import mediasoft.ru.backend.dto.response.product.CreateProductResponseDTO;
-import mediasoft.ru.backend.dto.response.product.ProductInfoResponseDTO;
-import mediasoft.ru.backend.dto.response.product.UpdateProductResponseDTO;
-import mediasoft.ru.backend.entities.Product;
+import mediasoft.ru.backend.models.dto.ProductDTO;
+import mediasoft.ru.backend.models.dto.response.product.CreateProductResponseDTO;
+import mediasoft.ru.backend.models.dto.response.product.ProductInfoResponseDTO;
+import mediasoft.ru.backend.models.dto.response.product.UpdateProductResponseDTO;
+import mediasoft.ru.backend.models.entities.Product;
 import mediasoft.ru.backend.enums.CurrencyEnum;
 import mediasoft.ru.backend.exceptions.ContentNotFoundException;
 import mediasoft.ru.backend.exceptions.EmptyFieldException;
 import mediasoft.ru.backend.exceptions.UniqueFieldException;
-import mediasoft.ru.backend.mappers.ProductMapper;
+import mediasoft.ru.backend.models.mappers.ProductMapper;
 import mediasoft.ru.backend.repositories.ProductRepository;
 import mediasoft.ru.backend.services.currency.CurrencyServiceClientImpl;
 import mediasoft.ru.backend.services.currency.CurrencyServiceMock;
