+++ backend/src/main/java/mediasoft/ru/backend/services/sheduling/DefaultSchedulerServiceImpl.java
@@ -1,8 +1,8 @@
-package mediasoft.ru.backend.sheduling;
+package mediasoft.ru.backend.services.sheduling;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import mediasoft.ru.backend.entities.Product;
+import mediasoft.ru.backend.models.entities.Product;
 import mediasoft.ru.backend.repositories.ProductRepository;
 import mediasoft.ru.backend.annotations.TimeMeasuring;
 import org.springframework.beans.factory.annotation.Value;
