+++ backend/src/main/java/mediasoft/ru/backend/services/sheduling/SchedulerService.java
@@ -1,4 +1,4 @@
-package mediasoft.ru.backend.sheduling;
+package mediasoft.ru.backend.services.sheduling;
 
 public interface SchedulerService {
     void raisePrice();
